Support GIF
1. Support GIF
2. Refact Renderer
3. Fix a bug in TextActionMode
4. Arrange code

+++ ukive/graphics/canvas.h
@@ -18,10 +18,9 @@ namespace ukive {
     class Window;
     class TextRenderer;
 
-    class Canvas
-    {
+    class Canvas {
     public:
-        Canvas(Window *win, float width, float height);
+        Canvas(Window* win, int width, int height);
         Canvas(ComPtr<ID2D1RenderTarget> renderTarget);
         ~Canvas();
 
@@ -35,17 +34,17 @@ namespace ukive {
         void endDraw();
 
         void popClip();
-        void pushClip(const RectF &rect);
+        void pushClip(const RectF& rect);
 
-        void pushLayer(ID2D1Geometry *clipGeometry);
-        void pushLayer(const RectF &content_bound, ID2D1Geometry *clipGeometry);
+        void pushLayer(ID2D1Geometry* clipGeometry);
+        void pushLayer(const RectF& content_bound, ID2D1Geometry* clipGeometry);
         void popLayer();
 
         void save();
         void restore();
 
         //临时方法。
-        ID2D1RenderTarget *getRT();
+        ID2D1RenderTarget* getRT();
 
         std::shared_ptr<Bitmap> extractBitmap();
 
@@ -55,55 +54,55 @@ namespace ukive {
         void rotate(float angle, float cx, float cy);
         void translate(float dx, float dy);
 
-        void setMatrix(const Matrix &matrix);
+        void setMatrix(const Matrix& matrix);
         Matrix getMatrix();
 
         void fillOpacityMask(
             float width, float height,
-            Bitmap *mask, Bitmap *content);
+            Bitmap* mask, Bitmap* content);
 
-        void drawRect(const RectF &rect, const Color &color);
-        void drawRect(const RectF &rect, float strokeWidth, const Color &color);
-        void fillRect(const RectF &rect, const Color &color);
+        void drawRect(const RectF& rect, const Color& color);
+        void drawRect(const RectF& rect, float strokeWidth, const Color& color);
+        void fillRect(const RectF& rect, const Color& color);
 
         void drawRoundRect(
-            const RectF &rect, float radius, const Color &color);
+            const RectF& rect, float radius, const Color& color);
         void drawRoundRect(
-            const RectF &rect, float strokeWidth,
-            float radius, const Color &color);
+            const RectF& rect, float strokeWidth,
+            float radius, const Color& color);
         void fillRoundRect(
-            const RectF &rect, float radius, const Color &color);
+            const RectF& rect, float radius, const Color& color);
 
-        void drawCircle(float cx, float cy, float radius, const Color &color);
-        void drawCircle(float cx, float cy, float radius, float strokeWidth, const Color &color);
-        void fillCircle(float cx, float cy, float radius, const Color &color);
+        void drawCircle(float cx, float cy, float radius, const Color& color);
+        void drawCircle(float cx, float cy, float radius, float strokeWidth, const Color& color);
+        void fillCircle(float cx, float cy, float radius, const Color& color);
 
-        void drawCircle(const RectF &rect, const Color &color);
-        void drawCircle(const RectF &rect, float strokeWidth, const Color &color);
-        void fillCircle(const RectF &rect, const Color &color);
+        void drawCircle(const RectF& rect, const Color& color);
+        void drawCircle(const RectF& rect, float strokeWidth, const Color& color);
+        void fillCircle(const RectF& rect, const Color& color);
 
-        void drawOval(float cx, float cy, float radiusX, float radiusY, const Color &color);
-        void drawOval(float cx, float cy, float radiusX, float radiusY, float strokeWidth, const Color &color);
-        void fillOval(float cx, float cy, float radiusX, float radiusY, const Color &color);
+        void drawOval(float cx, float cy, float radiusX, float radiusY, const Color& color);
+        void drawOval(float cx, float cy, float radiusX, float radiusY, float strokeWidth, const Color& color);
+        void fillOval(float cx, float cy, float radiusX, float radiusY, const Color& color);
 
-        void fillGeometry(ID2D1Geometry *geo, ID2D1Brush *brush);
+        void fillGeometry(ID2D1Geometry* geo, ID2D1Brush* brush);
 
-        void drawBitmap(Bitmap *bitmap);
-        void drawBitmap(float x, float y, Bitmap *bitmap);
-        void drawBitmap(const RectF &dst, float opacity, Bitmap *bitmap);
-        void drawBitmap(const RectF &src, const RectF &dst, float opacity, Bitmap *bitmap);
+        void drawBitmap(Bitmap* bitmap);
+        void drawBitmap(float x, float y, Bitmap* bitmap);
+        void drawBitmap(const RectF& dst, float opacity, Bitmap* bitmap);
+        void drawBitmap(const RectF& src, const RectF& dst, float opacity, Bitmap* bitmap);
 
         void drawText(
             std::wstring text,
-            IDWriteTextFormat *textFormat,
-            const RectF &layoutRect, const Color &color);
+            IDWriteTextFormat* textFormat,
+            const RectF& layoutRect, const Color& color);
         void drawTextLayout(
             float x, float y,
-            IDWriteTextLayout *textLayout, const Color &color);
+            IDWriteTextLayout* textLayout, const Color& color);
         void drawTextLayoutWithEffect(
-            View *v,
+            View* v,
             float x, float y,
-            IDWriteTextLayout *textLayout, const Color &color);
+            IDWriteTextLayout* textLayout, const Color& color);
 
     private:
         void initCanvas(ComPtr<ID2D1RenderTarget> renderTarget);
